pyproject.toml
[project]
name = "resume-critiquer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx>=0.2.4",
    "fpdf>=1.7.2",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.21",
    "langgraph>=0.4.8",
    "markdown>=3.8.1",
    "openai>=1.85.0",
    "pdfkit>=1.0.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "weasyprint>=65.1",
]
